fix: excludedPath not iterable
We saw an error in a user's deploy (helpdesk ticket: https://netlify.zendesk.com/agent/tickets/200248) that said:

```
Packaging Edge Functions from netlify/edge-functions directory:
  - __csp-nonce
 TypeError: params.excludedPath is not iterable
     at file:///opt/build/repo/netlify/edge-functions/__csp-nonce.ts:179:15
```

Looking at the code, the types for `params.excludedPath` say it can be a `string | Array<string>`. In case the property is a string, `...params.excludedPath` would simply make the string unusable for the config. For example:

```ts
const foo = 'bar'
console.log(...foo) // b a r
```

Since we need an array, I think it's best to just concat the 2 arrays (or an array and a string) together as they'd both produce the expected results.

Any thoughts?