perf: return early if the request is not a HTTP GET
This function should only be run its main logic if the request is a HTTP GET. By returning early instead of using `context.next(request)`, we avoid having the request go through this edge function, which will reduce the overall CPU and memory usage of this function.

Note that this behavior of returning `undefined` to bypass an Edge Function is documented behavior, it is safe to rely on this. It's documented on https://docs.netlify.com/edge-functions/api/:
```
The expected return value is one of the following:

    a standard Response object representing the HTTP response to be delivered to the client
    a standard URL object if you want to rewrite the incoming request to another same-site URL with a 200 status code
    undefined if you choose to bypass the current function
```