feat: Do not invoke the CSP Edge Function for Netlify Image CDN requests
Always include the `/.netlify/image` path in the `excludedPaths` array.

If the project is using NextJS, then also include `/_next/image` in the `excludedPaths` array. The reason this is only done for NextJS projects is because for projects not using NextJS, we would not know ahead of time what content-type this path would respond with, for example it could be responding with a 404 HTML page, which we would want to invoke the CSP Edge Function for.