fix: use entrypoint which embeds the wasm instead of requiring it to be fetched over the network when deployed as a Netlify Edge Function (#118)
Netlify Edge Functions are not able to read the file-system, which means we can't read the generated .wasm file during execution, we currently fetch the wasm file over the network but this request could fail or have an incorrect response, leading to a runtime error when attempting to instantiate the WebAssembly Module. To work-around this, the library was updated to include a .ts file which includes the exact same bytes as the wasm file, but within a JavaScript UInt8Array, which we can then instantiate as a WebAssembly Module.